+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/accident/Accident.java
@@ -1,6 +1,6 @@
 package kr.ac.kumoh.illdang100.tovalley.domain.accident;
 
-import kr.ac.kumoh.illdang100.tovalley.domain.valley.Valley;
+import kr.ac.kumoh.illdang100.tovalley.domain.waterplace.WaterPlace;
 import lombok.*;
 
 import javax.persistence.*;
@@ -19,8 +19,8 @@ public class Accident {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "valley_id")
-    private Valley valley;
+    @JoinColumn(name = "water_place_id")
+    private WaterPlace waterPlace;
 
     @Column(nullable = false)
     private LocalDate accidentDate;
