+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/review/Review.java
@@ -1,6 +1,6 @@
 package kr.ac.kumoh.illdang100.tovalley.domain.review;
 
-import kr.ac.kumoh.illdang100.tovalley.domain.valley.Valley;
+import kr.ac.kumoh.illdang100.tovalley.domain.waterplace.WaterPlace;
 import lombok.*;
 
 import javax.persistence.*;
@@ -18,8 +18,8 @@ public class Review {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "valley_id")
-    private Valley valley;
+    @JoinColumn(name = "water_place_id")
+    private WaterPlace waterPlace;
 
     @Column(nullable = false, length = 256)
     private String reviewContent;
