+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/trip_schedule/TripSchedule.java
@@ -1,6 +1,6 @@
 package kr.ac.kumoh.illdang100.tovalley.domain.trip_schedule;
 
-import kr.ac.kumoh.illdang100.tovalley.domain.valley.Valley;
+import kr.ac.kumoh.illdang100.tovalley.domain.waterplace.WaterPlace;
 import kr.ac.kumoh.illdang100.tovalley.domain.member.Member;
 import lombok.*;
 
@@ -24,8 +24,8 @@ public class TripSchedule {
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "valley_id")
-    private Valley valley;
+    @JoinColumn(name = "water_place_id")
+    private WaterPlace waterPlace;
 
     @Column(nullable = false)
     private LocalDate tripDate;
