--- tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/valley/ValleyRepository.java
@@ -1,6 +1,0 @@
-package kr.ac.kumoh.illdang100.tovalley.domain.valley;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ValleyRepository extends JpaRepository<Valley, Long> {
-}
