+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/waterplace/WaterPlaceRepository.java
@@ -1,0 +1,6 @@
+package kr.ac.kumoh.illdang100.tovalley.domain.waterplace;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface WaterPlaceRepository extends JpaRepository<WaterPlace, Long> {
+}
