--- tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/weather/valley_weather/ValleyWeatherRepository.java
@@ -1,6 +1,0 @@
-package kr.ac.kumoh.illdang100.tovalley.domain.weather.valley_weather;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ValleyWeatherRepository extends JpaRepository<ValleyWeather, Long> {
-}
