+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/weather/water_place_weather/WaterPlaceWeatherRepository.java
@@ -1,0 +1,6 @@
+package kr.ac.kumoh.illdang100.tovalley.domain.weather.water_place_weather;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface WaterPlaceWeatherRepository extends JpaRepository<WaterPlaceWeather, Long> {
+}
